% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{coerce_coef}
\alias{coerce_coef}
\title{Coerce coefficients}
\usage{
coerce_coef(p, as.type)
}
\arguments{
\item{p}{a polynomial}

\item{as.type}{the type coercion function}
}
\value{
a polynomial with coerced coefficients
}
\description{
Coerce the coefficients of a polynomial to a specific type.
}
