% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{common_eigenbasis}
\alias{common_eigenbasis}
\alias{common_eigenbasis0}
\title{Common eigenbasis}
\usage{
common_eigenbasis(As, tol = 1e-06, left = FALSE)

common_eigenbasis0(As, tol = 1e-06, left = FALSE)
}
\arguments{
\item{As}{a list of commuting matrices}

\item{tol}{tolerance used to determine if two eigenvalues differ.}

\item{left}{if \code{TRUE}, compute a common left eigenbasis}
}
\value{
a list of matrices, the columns of each matrix are the
common eigenbasis for the corresponding matrix in \code{As}
}
\description{
Compute a common eigenbasis for a list of commuting matrices.
}
\details{
Compute the common eigenbasis of a list of commuting matrices. No
check is performed to ensure the matrices are commute.

\code{common_eigenbasis} computes a common eigenbasis by computing the
eigenbasis of the first matrix and then rotating and splitting the
basis to align with the eigenbasis of each subsequent matrix.

\code{common_eigenbasis0} computes a common eigenbasis by computing the
eigenbasis of a random linear combination of the matrices.

No check is performed to ensure the matrices are commute.

If \code{tol} is too small, equal eigenvalues are erroneously identified
as distinct, and the bases for some eigenspaces are split.
}
