% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{derivative}
\alias{derivative}
\title{Derivative}
\usage{
derivative(p, order)
}
\arguments{
\item{p}{a polynomial}

\item{order}{vector of non-negative integers specifying the
order of the derivative}
}
\value{
the polynomial that is the derivative of \code{p}.
}
\description{
Compute the derivative of a polynomial.
}
