% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{derivative_matrix}
\alias{derivative_matrix}
\title{Derivative matrix}
\usage{
derivative_matrix(v, ms)
}
\arguments{
\item{v}{a variable index}

\item{ms}{a list of the exponent vectors of the monomial basis}
}
\value{
the derivative matrix
}
\description{
Compute the derivative matrix for a variable.
}
\details{
Computes the matrix representation in the basis \code{ms} of
differentiation by the \code{v}-th standard variable.
}
