% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{eval_poly}
\alias{eval_poly}
\alias{eval_polys}
\title{Evaluate polynomial}
\usage{
eval_poly(p, x)

eval_polys(ps, x)
}
\arguments{
\item{p}{a polynomial}

\item{x}{a vector or matrix of evaluation points}

\item{ps}{a list of polynomials}
}
\value{
the value of the polynomial at each x as complex or double.
}
\description{
Evaluate a polynomial or a list of polynomials at a point.
}
