% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner-package.R
\docType{package}
\name{groebner-package}
\alias{groebner-package}
\title{groebner: Groeber Bases}
\description{
Groebner Basis methods for solving systems of polynomials.
}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

}
\keyword{internal}
