% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{groebner}
\alias{groebner}
\title{Groebner basis}
\usage{
groebner(ps)
}
\arguments{
\item{ps}{a list of polynomials}
}
\value{
the Groebner basis
}
\description{
Compute the reduced Groebner basis of a set of polynomials
}
