% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{is_univariate}
\alias{is_univariate}
\title{polynomial is univariate}
\usage{
is_univariate(p)
}
\arguments{
\item{p}{a polynomial}
}
\value{
The index of the variable if p is univariate, 0 otherwise
}
\description{
Test if a polynomial is univariate.
}
