% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{macaulay_matrix}
\alias{macaulay_matrix}
\title{Macaulay matrix}
\usage{
macaulay_matrix(ps, tdeg = sum(degs - 1L) + 1L)
}
\arguments{
\item{ps}{a list of polynomials}

\item{tdeg}{the degree of the Macaulay matrix}
}
\value{
a list with elements
\itemize{
\item \code{M}: the Macaulay matrix
\item \code{ms}: the monomial basis
}
}
\description{
Compute the Macaulay matrix for a polynomial system.
}
