% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{macaulay_monomial_basis}
\alias{macaulay_monomial_basis}
\title{Macaulay monomial basis}
\usage{
macaulay_monomial_basis(ps)
}
\arguments{
\item{ps}{a list of polynomials}
}
\value{
a list with elements
\itemize{
\item \code{A}: reduction matrix
\item \code{term}: reducible terms
\item \code{basis}: the monomial basis
}
}
\description{
Compute a monomial basis determined from the dependencies of a
Macaulay matrix.
}
\details{
Computes a monomial basis from the linear dependences of the
Macaulay matrix of a polynomial system.  Returns a monomial basis
and a reduction matrix that can be used to reduce terms to the
monomial basis.
}
