% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{macaulay_multiplication_matrix}
\alias{macaulay_multiplication_matrix}
\title{Multiplication matrix from a Macaulay monomial basis}
\usage{
macaulay_multiplication_matrix(v, mb)
}
\arguments{
\item{v}{a variable index}

\item{mb}{an object returned by \code{\link[=macaulay_monomial_basis]{macaulay_monomial_basis()}}}
}
\value{
the multiplication matrix
}
\description{
Compute the multiplication matrix for a variable from a Macaulay
monomial basis.
}
