% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{monomial_basis}
\alias{monomial_basis}
\alias{monomial_basis0}
\title{Monomial basis}
\usage{
monomial_basis(gb)

monomial_basis0(rexpts)
}
\arguments{
\item{gb}{a Groebner basis}

\item{rexpts}{a list of reducible exponent vectors}
}
\value{
a list of the exponent vectors in the monomial basis
}
\description{
Compute a monomial basis from a Groebner basis or a list of
reducible monomials.
}
\details{
Given a Groebner basis or a list of reducible monomials terms
compute the corresponding monomial basis. The monomial basis is
returned as a list of integer vectors of the exponents of the of
the standard variables \verb{x1,x2,...} in the monomial term. The
\code{monomial_basis0} function computes a monomial basis given a list
of reducible terms, and \code{monomial_basis} computes a monomial basis
from a Groebner basis.
}
