% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{monomial_set}
\alias{monomial_set}
\title{Monomial set}
\usage{
monomial_set(nvar, deg)
}
\arguments{
\item{nvar}{the number of variables}

\item{deg}{the degree of the monomials}
}
\value{
a list of integer vectors representing the monomial set
}
\description{
Generate a set of monomials upto a given degree.
}
