% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{multiplication_matrix}
\alias{multiplication_matrix}
\title{Multiplication matrix}
\usage{
multiplication_matrix(v, ms, gb)
}
\arguments{
\item{v}{a variable index}

\item{ms}{a list of the exponent vectors of the monomial basis}

\item{gb}{a Groebner basis}
}
\value{
the multiplication matrix
}
\description{
Compute the multiplication matrix for a variable.
}
\details{
Computes the matrix representation in the basis \code{ms} of
multiplication by the \code{v}-th standard variable.
}
