% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{neval_poly}
\alias{neval_poly}
\alias{neval_polys}
\title{Evaluate polynomial}
\usage{
neval_poly(p, x)

neval_polys(ps, x)
}
\arguments{
\item{p}{a polynomial}

\item{x}{a vector or matrix of evaluation points}

\item{ps}{a list of polynomials}
}
\value{
the value of the polynomial at each x as complex or double.
}
\description{
Numerically evaluate a polynomial or a list of polynomials at an
array of points.
}
\details{
The function \code{neval_poly} evaluates a single polynomial at an array
of evaluation points returning the result as a numeric or complex
vector.  The function \code{neval_polys} evaluates a list of polynomials
at an array of evaluation points returning the result as a numeric
or complex matrix.
}
