% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{newton_polish}
\alias{newton_polish}
\title{Newton's method}
\usage{
newton_polish(ps, x, n)
}
\arguments{
\item{ps}{a list of polynomials}

\item{x}{a vector or matrix of initial estimates}

\item{n}{the number of iterations}
}
\value{
a root of the polynomial system
}
\description{
Polish a root of a polynomial system using Newton's method.
}
\details{
Given a root of a polynomial system, this function performs \code{n}
iterations of Newton's method to polish the root.
}
