% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(p)
}
\arguments{
\item{p}{a polynomial.}
}
\value{
The normalized polynomial.
}
\description{
Normalize a polynomial to have unit leading coefficient.
}
