% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{parse_poly}
\alias{parse_poly}
\alias{parse_polys}
\title{Polynomial parser}
\usage{
parse_poly(str, vars)

parse_polys(str, vars)
}
\arguments{
\item{str}{a character vector}

\item{vars}{a character vector of variable names}
}
\value{
a polynomial object
}
\description{
Parse string representations of polynomials
}
\details{
The \code{parse_poly} function converts a string representation of a
polynomial in the variables \code{vars} into a \code{poly} object. The
\code{parse_polys} function converts a list of strings into a list of
\code{poly} objects.

Note that the variables in \code{var} are translated to the standard
variables \verb{x1,x2,x3,...}.
}
\examples{
parse_poly("x^2*y^3-x*y*z",c("x","y","z"))
}
