% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{parse_term}
\alias{parse_term}
\title{Parse term}
\usage{
parse_term(str, vars)
}
\arguments{
\item{str}{a character vector}

\item{vars}{a character vector of variable names}
}
\value{
a polynomial term
}
\description{
Parse a string representation of a signed polynomial term.
}
\details{
Parses a signed polynomial term, returning the term and the
unconsumed component of the string.

Note that the variables in \code{var} are translated to the standard
variables \verb{x1,x2,x3,...}.
}
