% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{poly_add}
\alias{poly_add}
\alias{poly_sub}
\alias{poly_mul}
\alias{poly_scale}
\title{Polynomial arithmetic}
\usage{
poly_add(p1, p2)

poly_sub(p1, p2)

poly_mul(p1, p2)

poly_scale(acoef = 1L, aexpt = 0L, p1)
}
\arguments{
\item{p1, p2}{polynomials}

\item{acoef}{the coefficient of a}

\item{aexpt}{the exponent of a}
}
\value{
a polynomial
}
\description{
Polynomial arithmetic
}
\details{
\code{poly_add} computes p1+p2, \code{poly_sub} computes p1-p2, \code{poly_mul}
computes p1\emph{p2, and \code{poly_scale} computes a}p where a is a term.
}
