% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{poly_axpy}
\alias{poly_axpy}
\title{Polynomial axpy}
\usage{
poly_axpy(acoef = 1L, aexpt = 0L, px, py)
}
\arguments{
\item{acoef}{the coefficient of a}

\item{aexpt}{the exponent of a}

\item{px}{a polynomial}

\item{py}{a polynomial}
}
\value{
the polynomial a*px + py
}
\description{
Form the sum a*px + py where px and py are polynomials and a is a
monomial term.
}
