% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{poly_degree}
\alias{poly_degree}
\title{Degree of a polynomial}
\usage{
poly_degree(p)
}
\arguments{
\item{p}{a polynomial}
}
\value{
the degree of the polynomial
}
\description{
Compute the degree of a polynomial.
}
