% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{reorder_vars}
\alias{reorder_vars}
\title{Reorder variables}
\usage{
reorder_vars(p, order)
}
\arguments{
\item{p}{a polynomial}

\item{order}{an integer vector defining the new variable order}
}
\value{
a polynomial with the variables in the new order
}
\description{
Change the variable order of a polynomial
}
\details{
Permute the variables in a polynomial according to a permutation
vector \code{order}.  For example, if \code{order} is \code{c(3,2,1)},then the
standard variables \code{x1} and \code{x3} are interchanged and the terms
are reordered according to the term order.

If \code{order} specifies more or fewer variables than are present in
the current variable set, the exponent vectors are resized to
encode the new variable set.
}
