% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{roots}
\alias{roots}
\title{Roots from eigenvalues and eigenvectors}
\usage{
roots(Ms, Bs, unique = TRUE)
}
\arguments{
\item{Ms}{a list of multiplication matrices}

\item{Bs}{an eigenbasis}

\item{unique}{In degenerate cases, return each root only once}
}
\value{
a matrix where each row is a root of the polynomial system
}
\description{
Extract the roots of a polynomial system
}
\details{
\code{roots} computes the roots of a polynomial system from the
eigenvalues of its multiplication matrices using a common basis
computed by \code{\link[=common_eigenbasis]{common_eigenbasis()}}.  The roots are returned in the
same order as the multiplication matrices.
}
