% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{roots_eigenvectors}
\alias{roots_eigenvectors}
\title{Roots from eigenvectors}
\usage{
roots_eigenvectors(ms, Bs)
}
\arguments{
\item{ms}{a monomial basis}

\item{Bs}{a common eigenbasis}
}
\value{
a matrix where each row is a root of the polynomial system
}
\description{
Extract the roots of a polynomial system from the right eigenvectors
of the multiplication matrices.
}
\details{
If \code{Bs} is a common right eigenbasis calculated from the transposed
multiplication matrices, then this function will extract the values
of the standard variables that appear in the monomials in the basis
\code{ms} at the roots of the polynomial system.
}
