% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{solve_polys}
\alias{solve_polys}
\title{Roots of a polynomial system}
\usage{
solve_polys(ps, gb = groebner(ps), newton = 0, tol = 1e-07)
}
\arguments{
\item{ps}{a list of polynomials}

\item{gb}{a Groebner basis}

\item{newton}{the number of Newton iterations used to refine roots}

\item{tol}{tolerance used to determine if two eigenvalues differ.}
}
\value{
a matrix where each row is a root of the polynomial system
}
\description{
Find the roots of a polynomial system
}
