% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{sort_list}
\alias{sort_list}
\title{Sort list}
\usage{
sort_list(lst, cmp)
}
\arguments{
\item{lst}{a list}

\item{cmp}{comparison function}
}
\value{
the sorted list
}
\description{
Sort the elements of a list.
}
\details{
Sort the elements of a list according to a comparison function
\code{cmp(x,y)} which returns
\itemize{
\item -1 if \code{x < y}
\item 0 if \code{x==y}
\item 1 if \code{x > y}.
}
}
