% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{spoly}
\alias{spoly}
\title{S-polynomial}
\usage{
spoly(p1, p2)
}
\arguments{
\item{p1}{a polynomial}

\item{p2}{a polynomial}
}
\value{
the S-polynomial
}
\description{
Compute the S-polynomial of two polynomials
}
