% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{lex}
\alias{lex}
\alias{grlex}
\alias{grevlex}
\alias{set_term_order}
\alias{get_term_order}
\title{Term Orders}
\usage{
lex(expt1, expt2)

grlex(expt1, expt2)

grevlex(expt1, expt2)

set_term_order(order = grevlex)

get_term_order(order = grevlex)
}
\arguments{
\item{expt1}{integer vector of variable exponents}

\item{expt2}{integer vector of variable exponents}

\item{order}{a term order function}
}
\value{
The term order functions return
\itemize{
\item -1 if \code{expt1 < expt2}
\item 0 if \code{expt1 == expt2}
\item 1 if \code{expt1 > expt2}
}
}
\description{
Polynomial term orders.
}
\details{
Three standard term ordering are available
\itemize{
\item \code{lex} lexicographic
\item \code{grlex} graded lexicographic
\item \code{grevlex} reverse graded lexicographic
}

but any arbitrary term order can be used. The default term order
is \code{grevlex}.

The current term can be set and retrieved with the set_term_order
and get_term_order functions.

If the term order is changed, any existing polynomials are left in
an inconsistent state, and it is necessary to reapply \code{\link[=poly]{poly()}} to
recover a valid polynomial.
}
\seealso{
\code{\link[=reorder_vars]{reorder_vars()}}
}
