% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{is_univariate}
\alias{is_univariate}
\alias{as_univariate}
\title{Univariate polynomial}
\usage{
is_univariate(p)

as_univariate(p, v)
}
\arguments{
\item{p}{a polynomial.}

\item{v}{the variable index.}
}
\value{
\code{is_univariate} returns the index of the variable if p is
univariate, 0 otherwise. \code{as_univariate} returns the vector of
coefficients of a univariate polynomial.
}
\description{
Test if a polynomial is univariate and extract the vector of
coefficients.
}
\details{
The function \code{is_univariate} tests if \code{p} is a polynomial in a
single variable, returning the index of the variable if it is and
zero otherwise. The function \code{as_univariate} extracts the vector
of coefficients of a polynomial that is univariate in the variable
with index \code{v}.
}
