% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{eigenbasis}
\alias{eigenbasis}
\title{Eigenbasis}
\usage{
eigenbasis(A, tol = 1e-06)
}
\arguments{
\item{A}{a matrix}

\item{tol}{tolerance used to determine if two eigenvalues differ.}
}
\value{
a list of matrices, one for each distinct eigenvalue, the
columns of each are the eigenbasis for that eigenvalue
}
\description{
Compute an eigenbasis of a matrix.
}
