% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{gauss_jordan}
\alias{gauss_jordan}
\title{Gauss-Jordan elimination}
\usage{
gauss_jordan(A)
}
\arguments{
\item{A}{a matrix}
}
\value{
a list with elements
\itemize{
\item \code{A}: the reduced matrix
\item \code{reduced}: the indices of the reduced columns
\item \code{rank}: the rank of the matrix
}
}
\description{
reduce a matrix to reduced row echelon form by Gauss-Jordan
elimination.
}
