% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{monomial_basis}
\alias{monomial_basis}
\title{Monomial basis}
\usage{
monomial_basis(gb)
}
\arguments{
\item{gb}{a Groebner basis}
}
\value{
a list of integer vectors representing the monomial basis
}
\description{
Compute a monomial basis.
}
\details{
Given a groebner basis, compute the corresponding monomial basis.
The monomial basis is returned as a list of integer vectors each
representing the powers of the standard variables \verb{x1,x2,...} in
the monomial term.
}
