% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{pterm}
\alias{pterm}
\title{Create a monomial}
\usage{
pterm(coef, expt)
}
\arguments{
\item{coef}{the coefficient of the term}

\item{expt}{integer vector of variable exponents}
}
\value{
a pterm object
}
\description{
Create  monomial term
}
\details{
A monomial term is represented as a coefficient and an integer
vector of the exponents of the standard variables \verb{x1,x2,...} in
the monomial term. This function is used in conjuction with
\code{\link[=poly]{poly()}} to create a polynomial.
}
\seealso{
\code{\link[=poly]{poly()}}
}
