% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groebner.R
\name{reduce_pp}
\alias{reduce_pp}
\alias{reduce_ps}
\alias{reduce_sp}
\alias{reduce_s}
\title{Reduce}
\usage{
reduce_pp(p1, p2)

reduce_ps(p, ps)

reduce_sp(ps, p)

reduce_s(ps)
}
\arguments{
\item{p1}{a polynomial}

\item{p2}{a polynomial}

\item{p}{a polynomial}

\item{ps}{a list of polynomials}
}
\value{
the reduced polynomial
}
\description{
Reduce polynomials.
}
\details{
There are four functions for reducing polynomials:
\itemize{
\item \code{reduce_pp} reduces a polynomial by another polynomial
\item \code{reduce_ps} reduces a polynomial by a list of polynomials
\item \code{reduce_sp} reduces a list of polynomials by a polynomial
\item \code{reduce_s} reduces a list of polynomials
}

For functions \code{reduce_pp} and \code{reduce_sp} the reductions are
performed in a single pass, but for \code{reduce_ps} and \code{reduce_s} the
reductions are iterated until the result cannot be further
reduced.
}
